Disable unused conditional checks
Where we are not using some options, I have disabled checking for them, hoping to get a minor performance boost. This commit also fixes a bit of test code that was left in. The Document class will now get features from the `features` property if it is already present in the manifest.